Creates the advice model and proves its use interface with simple unit tests
I usually start with a new project by its "most important flow", in this case, it is the
list advices flow. Inside the flow, I usually start by the model, I usually create the
simplest model as possible followed by some unit tests, which support me to understand how
the model interface will look like and how pleasant/or not it is to use that model's interface.

Next step, create the use case to implement the flow.