Moves the model classes like Advice and Type to the domain.model package.
This commit also moves the Type from the Advice class to its own file, creates
an "of" method in the enum which converts a string into the enum or returns
Optional.empty() when the string does not represent one of the enum values.

Unit tests for the Type conversion were also added in this commit.

Next step, use case implementation...