Implements the show advices use case and its unit tests.
The use case uses a callback, similar to what was used in the past in Android
development, the called "Delegates". Even though the code may look a bit verbose,
I like to make the possible returns explicit in the use case via the callbacks.

In the unit tests, I used mocks for the callback but not for the repository, once
mocks may introduce "behaviour drifts" once the real code implementation evolves. On
this topic I like the article from Google
https://testing.googleblog.com/2013/05/testing-on-toilet-dont-overuse-mocks.html.